+++ pukiwiki/plugin/freeze.inc.php
@@ -28,7 +28,7 @@ function plugin_freeze_action()
 		$body = str_replace('$1', htmlsc(strip_bracket($page)),
 			$_title_isfreezed);
 
-	} else if ($pass !== NULL && pkwk_login($pass)) {
+	} else if (isset($vars['ok']) && pkwk_admin_authorized($pass)) {
 		// Freeze
 		$postdata = get_source($page);
 		array_unshift($postdata, "#freeze\n");
@@ -44,14 +44,17 @@ function plugin_freeze_action()
 		// Show a freeze form
 		$msg    = & $_title_freeze;
 		$s_page = htmlsc($page);
-		$body   = ($pass === NULL) ? '' : "<p><strong>$_msg_invalidpass</strong></p>\n";
+		$body   = (isset($vars['ok']) && ! pkwk_admin_authorized($pass)) ?
+			"<p><strong>$_msg_invalidpass</strong></p>\n" : '';
+		$pass_input = pkwk_is_authenticated() ? '' :
+			'<input type="password" name="pass" size="12" />';
 		$body  .= <<<EOD
 <p>$_msg_freezing</p>
 <form action="$script" method="post">
  <div>
   <input type="hidden"   name="cmd"  value="freeze" />
   <input type="hidden"   name="page" value="$s_page" />
-  <input type="password" name="pass" size="12" />
+  $pass_input
   <input type="submit"   name="ok"   value="$_btn_freeze" />
  </div>
 </form>